+++ docs/source/installation.rst
@@ -19,9 +19,9 @@ All platforms
    For Linux" and use the Bash console to continue the instructions.
 
 1. Install Anaconda/Miniconda (https://docs.conda.io/en/latest/miniconda.html)
-2. Create new environment via :code:`conda create -n aligner -c conda-forge python=3.8 openfst=1.7.6 pynini=2.1.0 ngram=1.3.9 baumwelch=0.3.1`
+2. Create new environment via :code:`conda create -n aligner -c conda-forge openblas python=3.8 openfst=1.7.6 pynini=2.1.0 ngram=1.3.9 baumwelch=0.3.1`
 3. Run :code:`pip install montreal-forced-aligner`
-4. Install third-party binaries via :code:`mfa thirdparty download` (see also :ref:`collect_binaries` to collect locally built binaries)
+5. Install third-party binaries via :code:`mfa thirdparty download` (see also :ref:`collect_binaries` to collect locally built binaries)
 
 .. _collect_binaries:
