+++ montreal_forced_aligner/command_line/train_ivector_extractor.py
@@ -5,10 +5,14 @@ import yaml
 import time
 
 from montreal_forced_aligner import __version__
-from montreal_forced_aligner.corpus import TranscribeCorpus
-from montreal_forced_aligner.aligner import TrainableAligner
+from montreal_forced_aligner.corpus import AlignableCorpus
+from montreal_forced_aligner.dictionary import Dictionary
+from montreal_forced_aligner.aligner import PretrainedAligner
 from montreal_forced_aligner.config import TEMP_DIR, train_yaml_to_config, load_basic_train_ivector
 from montreal_forced_aligner.helper import setup_logger
+from montreal_forced_aligner.utils import get_available_acoustic_languages, get_pretrained_acoustic_path, \
+    get_available_dict_languages, get_dictionary_path
+from montreal_forced_aligner.models import AcousticModel
 
 from montreal_forced_aligner.exceptions import ArgumentError
 
@@ -43,10 +47,15 @@ def train_ivector(args):
                 'begin': all_begin,
                 'version': __version__,
                 'type': command,
-                'corpus_directory': args.corpus_directory}
+                'corpus_directory': args.corpus_directory,
+                'dictionary_path': args.dictionary_path,
+                'acoustic_model_path': args.acoustic_model_path,
+                }
     if conf['dirty'] or conf['type'] != command \
             or conf['corpus_directory'] != args.corpus_directory \
-            or conf['version'] != __version__:
+            or conf['version'] != __version__ \
+            or conf['dictionary_path'] != args.dictionary_path \
+            or conf['acoustic_model_path'] != args.acoustic_model_path:
         logger.warning(
             'WARNING: Using old temp directory, this might not be ideal for you, use the --clean flag to ensure no '
             'weird behavior for previous versions of the temporary directory.')
@@ -59,21 +68,40 @@ def train_ivector(args):
                          'path {} (new run: {})'.format(conf['corpus_directory'], args.corpus_directory))
         if conf['version'] != __version__:
             logger.debug('Previous run was on {} version (new run: {})'.format(conf['version'], __version__))
+        if conf['dictionary_path'] != args.dictionary_path:
+            logger.debug('Previous run used dictionary path {} '
+                         '(new run: {})'.format(conf['dictionary_path'], args.dictionary_path))
+        if conf['acoustic_model_path'] != args.acoustic_model_path:
+            logger.debug('Previous run used acoustic model path {} '
+                         '(new run: {})'.format(conf['acoustic_model_path'], args.acoustic_model_path))
 
     os.makedirs(data_directory, exist_ok=True)
     try:
-        corpus = TranscribeCorpus(args.corpus_directory, data_directory,
-                                  speaker_characters=args.speaker_characters,
-                                  num_jobs=args.num_jobs,
-                             debug=getattr(args, 'debug', False), logger=logger)
-        dictionary = None
-        a = TrainableAligner(corpus, dictionary, train_config, align_config,
-                             temp_directory=data_directory, logger=logger)
+        begin = time.time()
+        corpus = AlignableCorpus(args.corpus_directory, data_directory,
+                                 speaker_characters=args.speaker_characters,
+                                 num_jobs=args.num_jobs,
+                                 debug=getattr(args, 'debug', False), logger=logger)
+        acoustic_model = AcousticModel(args.acoustic_model_path)
+        dictionary = Dictionary(args.dictionary_path, data_directory, word_set=corpus.word_set, logger=logger)
+        acoustic_model.validate(dictionary)
+        a = PretrainedAligner(corpus, dictionary, acoustic_model, align_config,
+                              temp_directory=data_directory, logger=logger)
+        logger.debug('Setup pretrained aligner in {} seconds'.format(time.time() - begin))
         a.verbose = args.verbose
         begin = time.time()
-        a.train()
-        logger.debug('Training took {} seconds'.format(time.time() - begin))
-        a.save(args.output_model_path)
+        a.align()
+        logger.debug('Performed alignment in {} seconds'.format(time.time() - begin))
+        for identifier, trainer in train_config.items():
+            trainer.logger = logger
+            if identifier != 'ivector':
+                continue
+            begin = time.time()
+            trainer.init_training(identifier, data_directory, corpus, dictionary, a)
+            trainer.train(call_back=print)
+            logger.debug('Training took {} seconds'.format(time.time() - begin))
+            trainer.save(args.output_model_path)
+
         logger.info('All done!')
         logger.debug('Done! Everything took {} seconds'.format(time.time() - all_begin))
     except Exception as e:
@@ -84,7 +112,7 @@ def train_ivector(args):
             yaml.dump(conf, f)
 
 
-def validate_args(args):
+def validate_args(args, downloaded_acoustic_models, download_dictionaries):
     if args.config_path and not os.path.exists(args.config_path):
         raise (ArgumentError('Could not find the config file {}.'.format(args.config_path)))
 
@@ -93,24 +121,47 @@ def validate_args(args):
     if not os.path.isdir(args.corpus_directory):
         raise (ArgumentError('The specified corpus directory ({}) is not a directory.'.format(args.corpus_directory)))
 
+    if args.dictionary_path.lower() in download_dictionaries:
+        args.dictionary_path = get_dictionary_path(args.dictionary_path.lower())
+    if not os.path.exists(args.dictionary_path):
+        raise ArgumentError('Could not find the dictionary file {}'.format(args.dictionary_path))
+    if not os.path.isfile(args.dictionary_path):
+        raise ArgumentError('The specified dictionary path ({}) is not a text file.'.format(args.dictionary_path))
 
-def run_train_ivector_extractor(args):
+    if args.acoustic_model_path.lower() in downloaded_acoustic_models:
+        args.acoustic_model_path = get_pretrained_acoustic_path(args.acoustic_model_path.lower())
+    elif args.acoustic_model_path.lower().endswith(AcousticModel.extension):
+        if not os.path.exists(args.acoustic_model_path):
+            raise ArgumentError('The specified model path does not exist: ' + args.acoustic_model_path)
+    else:
+        raise ArgumentError(
+            'The language \'{}\' is not currently included in the distribution, '
+            'please align via training or specify one of the following language names: {}.'.format(
+                args.acoustic_model_path.lower(), ', '.join(downloaded_acoustic_models)))
+
+
+def run_train_ivector_extractor(args, downloaded_acoustic_models=None, download_dictionaries=None):
+    if downloaded_acoustic_models is None:
+        downloaded_acoustic_models = get_available_acoustic_languages()
+    if download_dictionaries is None:
+        download_dictionaries = get_available_dict_languages()
     try:
         args.speaker_characters = int(args.speaker_characters)
     except ValueError:
         pass
     args.corpus_directory = args.corpus_directory.rstrip('/').rstrip('\\')
 
-    validate_args(args)
+    validate_args(args, downloaded_acoustic_models, download_dictionaries)
     train_ivector(args)
 
 
 if __name__ == '__main__':  # pragma: no cover
     mp.freeze_support()
-    from montreal_forced_aligner.command_line.mfa import train_ivector_parser, fix_path, unfix_path
+    from montreal_forced_aligner.command_line.mfa import train_ivector_parser, fix_path, unfix_path, acoustic_languages, \
+        dict_languages
 
     ivector_args = train_ivector_parser.parse_args()
 
     fix_path()
-    run_train_ivector_extractor(ivector_args)
+    run_train_ivector_extractor(ivector_args, acoustic_languages, dict_languages)
     unfix_path()
