+++ montreal_forced_aligner/trainers/ivector_extractor.py
@@ -2,12 +2,16 @@ import os
 from tqdm import tqdm
 import subprocess
 import shutil
+import numpy as np
+from sklearn.naive_bayes import GaussianNB
+from joblib import dump, load
 
 from .base import BaseTrainer
-from ..helper import thirdparty_binary, make_path_safe, log_kaldi_errors, parse_logs
+from ..helper import thirdparty_binary, make_path_safe, log_kaldi_errors, parse_logs, load_scp
 from ..exceptions import KaldiProcessingError
 
-from ..multiprocessing import (gmm_gselect, acc_global_stats, gauss_to_post, acc_ivector_stats, extract_ivectors)
+from ..multiprocessing.ivector import (gmm_gselect, acc_global_stats, gauss_to_post,
+                                           acc_ivector_stats, extract_ivectors)
 
 from ..models import IvectorExtractor
 
@@ -64,6 +68,9 @@ class IvectorExtractorTrainer(BaseTrainer):
         self.subsample = 5
         self.max_count = 100
         self.apply_cmn = True
+        self.pca_dimension = -1
+        self.use_vad = False
+        self.previous_align_directory = None
 
     @property
     def meta(self):
@@ -94,17 +101,23 @@ class IvectorExtractorTrainer(BaseTrainer):
         if call_back is None:
             call_back = print
         # train diag ubm
+        final_ubm_path = os.path.join(self.train_directory, 'final.ubm')
+        if os.path.exists(final_ubm_path):
+            return
         self.logger.info('Initializing diagonal UBM...')
         # Initialize model from E-M in memory
         log_directory = os.path.join(self.train_directory, 'log')
         num_gauss_init = int(self.ubm_initial_gaussian_proportion * int(self.ubm_num_gaussians))
         log_path = os.path.join(log_directory, 'gmm_init.log')
-
-        all_feats_path = os.path.join(self.corpus.output_directory, self.feature_config.voiced_feature_id + '.scp')
+        if self.use_vad:
+            feat_name = self.feature_config.voiced_feature_id
+        else:
+            feat_name = self.feature_file_base_name
+        all_feats_path = os.path.join(self.corpus.output_directory, feat_name + '.scp')
         with open(all_feats_path, 'w') as outf:
             for i in range(self.corpus.num_jobs):
                 with open(os.path.join(self.data_directory,
-                                       self.feature_config.voiced_feature_id + '.{}.scp'.format(i))) as inf:
+                                       feat_name + '.{}.scp'.format(i))) as inf:
                     for line in inf:
                         outf.write(line)
         with open(log_path, 'w') as log_file:
@@ -120,130 +133,162 @@ class IvectorExtractorTrainer(BaseTrainer):
             gmm_init_proc.communicate()
 
         # Store Gaussian selection indices on disk
-        gmm_gselect('0', self, self.corpus.num_jobs)
+        gmm_gselect('0', self, self.corpus.num_jobs, vad=self.use_vad)
+        final_dubm_path = os.path.join(self.train_directory, 'final.dubm')
 
-        self.logger.info('Training diagonal UBM...')
-        if call_back == print:
-            iters = tqdm(range(0, self.ubm_num_iterations))
-        else:
-            iters = range(0, self.ubm_num_iterations)
-        for i in iters:
-            # Accumulate stats
-            acc_global_stats(self, self.corpus.num_jobs, i, full=False)
-
-            # Don't remove low-count Gaussians till the last tier,
-            # or gselect info won't be valid anymore
-            if i < self.ubm_num_iterations - 1:
-                opt = '--remove-low-count-gaussians=false'
+        if not os.path.exists(final_dubm_path):
+            self.logger.info('Training diagonal UBM...')
+            if call_back == print:
+                iters = tqdm(range(0, self.ubm_num_iterations))
             else:
-                opt = '--remove-low-count-gaussians={}'.format(self.ubm_remove_low_count_gaussians)
+                iters = range(0, self.ubm_num_iterations)
+            for i in iters:
+                # Accumulate stats
+                acc_global_stats(self, self.corpus.num_jobs, i, full=False, vad=self.use_vad)
 
-            log_path = os.path.join(self.train_directory, 'log', 'update.{}.log'.format(i))
+                # Don't remove low-count Gaussians till the last tier,
+                # or gselect info won't be valid anymore
+                if i < self.ubm_num_iterations - 1:
+                    opt = '--remove-low-count-gaussians=false'
+                else:
+                    opt = '--remove-low-count-gaussians={}'.format(self.ubm_remove_low_count_gaussians)
+
+                log_path = os.path.join(self.train_directory, 'log', 'update.{}.log'.format(i))
+                with open(log_path, 'w') as log_file:
+                    acc_files = [os.path.join(self.train_directory, '{}.{}.acc'.format(i, x))
+                                 for x in range(self.corpus.num_jobs)]
+                    gmm_global_est_proc = subprocess.Popen([thirdparty_binary('gmm-global-est'),
+                                                            opt,
+                                                            '--min-gaussian-weight=' + str(self.ubm_min_gaussian_weight),
+                                                            os.path.join(self.train_directory, '{}.dubm'.format(i)),
+                                                            "{} - {}|".format(thirdparty_binary('gmm-global-sum-accs'),
+                                                                              ' '.join(map(make_path_safe, acc_files))),
+                                                            os.path.join(self.train_directory, '{}.dubm'.format(i + 1))],
+                                                           stderr=log_file)
+                    gmm_global_est_proc.communicate()
+                    # Clean up
+                    for p in acc_files:
+                        os.remove(p)
+
+                # Move files
+            shutil.copy(os.path.join(self.train_directory, '{}.dubm'.format(self.ubm_num_iterations)),
+                        final_dubm_path)
+        if self.use_vad:
+            # train full ubm
+
+            # Convert final.ubm to fgmm
+            log_path = os.path.join(self.train_directory, 'log', 'global_to_fgmm.log')
             with open(log_path, 'w') as log_file:
-                acc_files = [os.path.join(self.train_directory, '{}.{}.acc'.format(i, x))
-                             for x in range(self.corpus.num_jobs)]
-                gmm_global_est_proc = subprocess.Popen([thirdparty_binary('gmm-global-est'),
-                                                        opt,
-                                                        '--min-gaussian-weight=' + str(self.ubm_min_gaussian_weight),
-                                                        os.path.join(self.train_directory, '{}.dubm'.format(i)),
-                                                        "{} - {}|".format(thirdparty_binary('gmm-global-sum-accs'),
-                                                                          ' '.join(map(make_path_safe, acc_files))),
-                                                        os.path.join(self.train_directory, '{}.dubm'.format(i + 1))],
-                                                       stderr=log_file)
-                gmm_global_est_proc.communicate()
+                subprocess.call([thirdparty_binary('gmm-global-to-fgmm'),
+                                 os.path.join(self.train_directory, 'final.dubm'),
+                                 os.path.join(self.train_directory, '0.ubm')],
+                                stdout=subprocess.PIPE,
+                                stderr=log_file)
+            gmm_gselect('final', self, self.corpus.num_jobs, vad=self.use_vad)
 
-            # Move files
-        shutil.copy(os.path.join(self.train_directory, '{}.dubm'.format(self.ubm_num_iterations)),
-                    os.path.join(self.train_directory, 'final.dubm'))
-
-        # train full ubm
-
-        # Convert final.ubm to fgmm
-        log_path = os.path.join(self.train_directory, 'log', 'global_to_fgmm.log')
-        with open(log_path, 'w') as log_file:
-            subprocess.call([thirdparty_binary('gmm-global-to-fgmm'),
-                             os.path.join(self.train_directory, 'final.dubm'),
-                             os.path.join(self.train_directory, '0.ubm')],
-                            stdout=subprocess.PIPE,
-                            stderr=log_file)
-        gmm_gselect('final', self, self.corpus.num_jobs)
-
-        self.logger.info('Training full UBM...')
-        if call_back == print:
-            iters = tqdm(range(0, self.ubm_num_iterations))
-        else:
-            iters = range(0, self.ubm_num_iterations)
-        for i in iters:
-            # Accumulate stats
-            acc_global_stats(self, self.corpus.num_jobs, i, full=True)
-
-            # Don't remove low-count Gaussians till the last tier,
-            # or gselect info won't be valid anymore
-            if i < self.ubm_num_iterations - 1:
-                opt = '--remove-low-count-gaussians=false'
+            self.logger.info('Training full UBM...')
+            if call_back == print:
+                iters = tqdm(range(0, self.ubm_num_iterations))
             else:
-                opt = '--remove-low-count-gaussians={}'.format(self.ubm_remove_low_count_gaussians)
+                iters = range(0, self.ubm_num_iterations)
+            for i in iters:
+                # Accumulate stats
+                acc_global_stats(self, self.corpus.num_jobs, i, full=True, vad=self.use_vad)
 
-            log_path = os.path.join(self.train_directory, 'log', 'update.{}.log'.format(i))
-            with open(log_path, 'w') as log_file:
-                acc_files = [os.path.join(self.train_directory, '{}.{}.acc'.format(i, x))
-                             for x in range(self.corpus.num_jobs)]
-                gmm_global_est_proc = subprocess.Popen([thirdparty_binary('fgmm-global-est'),
-                                                        opt,
-                                                        '--min-gaussian-weight=' + str(self.ubm_min_gaussian_weight),
-                                                        os.path.join(self.train_directory, '{}.ubm'.format(i)),
-                                                        "{} - {}|".format(thirdparty_binary('fgmm-global-sum-accs'),
-                                                                          ' '.join(map(make_path_safe, acc_files))),
-                                                        os.path.join(self.train_directory, '{}.ubm'.format(i + 1))],
-                                                       stderr=log_file)
-                gmm_global_est_proc.communicate()
-            # Move files
-        shutil.copy(os.path.join(self.train_directory, '{}.ubm'.format(self.ubm_num_iterations)),
-                    os.path.join(self.train_directory, 'final.ubm'))
+                # Don't remove low-count Gaussians till the last tier,
+                # or gselect info won't be valid anymore
+                if i < self.ubm_num_iterations - 1:
+                    opt = '--remove-low-count-gaussians=false'
+                else:
+                    opt = '--remove-low-count-gaussians={}'.format(self.ubm_remove_low_count_gaussians)
+
+                log_path = os.path.join(self.train_directory, 'log', 'update.{}.log'.format(i))
+                with open(log_path, 'w') as log_file:
+                    acc_files = [os.path.join(self.train_directory, '{}.{}.acc'.format(i, x))
+                                 for x in range(self.corpus.num_jobs)]
+                    gmm_global_est_proc = subprocess.Popen([thirdparty_binary('fgmm-global-est'),
+                                                            opt,
+                                                            '--min-gaussian-weight=' + str(self.ubm_min_gaussian_weight),
+                                                            os.path.join(self.train_directory, '{}.ubm'.format(i)),
+                                                            "{} - {}|".format(thirdparty_binary('fgmm-global-sum-accs'),
+                                                                              ' '.join(map(make_path_safe, acc_files))),
+                                                            os.path.join(self.train_directory, '{}.ubm'.format(i + 1))],
+                                                           stderr=log_file)
+                    gmm_global_est_proc.communicate()
+                    # Clean up
+                    for p in acc_files:
+                        os.remove(p)
+                # Move files
+            shutil.copy(os.path.join(self.train_directory, '{}.ubm'.format(self.ubm_num_iterations)),
+                        final_ubm_path)
         parse_logs(log_directory)
         self.logger.info('Finished training UBM!')
 
-    def init_training(self, identifier, temporary_directory, corpus, dictionary, previous_trainer):
+    def init_training(self, identifier, temporary_directory, corpus, dictionary, previous_trainer=None):
         self._setup_for_init(identifier, temporary_directory, corpus, dictionary)
         done_path = os.path.join(self.train_directory, 'done')
         dirty_path = os.path.join(self.train_directory, 'dirty')
         if os.path.exists(done_path):
             self.logger.info('{} training already done, skipping initialization.'.format(self.identifier))
             return
-        try:
-            self.feature_config.compute_vad(self.corpus, logger=self.logger)
-            self.feature_config.generate_voiced_features(self.corpus, apply_cmn=self.apply_cmn, logger=self.logger)
+        if previous_trainer is None:
+            self.use_vad = True
+        else:
+            self.previous_align_directory = previous_trainer.align_directory
+        if self.use_vad:
+            try:
+                self.feature_config.compute_vad(self.corpus, logger=self.logger)
+                self.feature_config.generate_voiced_features(self.corpus, apply_cmn=self.apply_cmn, logger=self.logger)
+                self.corpus.create_vad_segments(self.feature_config)
+                # Subsegment existing segments from TextGrids or VAD
+                self.corpus.create_subsegments(self.feature_config)
+            except Exception as e:
+                with open(dirty_path, 'w'):
+                    pass
+                if isinstance(e, KaldiProcessingError):
+                    log_kaldi_errors(e.error_logs, self.logger)
+                raise
 
-            self.train_ubm()
-            self.init_ivector_train()
-        except Exception as e:
-            with open(dirty_path, 'w'):
-                pass
-            if isinstance(e, KaldiProcessingError):
-                log_kaldi_errors(e.error_logs, self.logger)
-            raise
+        self.train_ubm()
+        self.init_ivector_train()
 
     def init_ivector_train(self):
-
+        init_ie_path = os.path.join(self.train_directory, '0.ie')
+        if os.path.exists(init_ie_path):
+            return
         # Initialize i-vector extractor
         log_directory = os.path.join(self.train_directory, 'log')
         log_path = os.path.join(log_directory, 'init.log')
         diag_ubm_path = os.path.join(self.train_directory, 'final.dubm')
         full_ubm_path = os.path.join(self.train_directory, 'final.ubm')
         with open(log_path, 'w') as log_file:
-            subprocess.call([thirdparty_binary('fgmm-global-to-gmm'),
-                             full_ubm_path,
-                             diag_ubm_path],
-                            stderr=log_file)
-            subprocess.call([thirdparty_binary('ivector-extractor-init'),
-                             '--ivector-dim=' + str(self.ivector_dimension),
-                             '--use-weights={}'.format(self.use_weights).lower(),
-                             full_ubm_path,
-                             os.path.join(self.train_directory, '0.ie')],
-                            stderr=log_file)
+            if self.use_vad:
+                subprocess.call([thirdparty_binary('fgmm-global-to-gmm'),
+                                 full_ubm_path,
+                                 diag_ubm_path],
+                                stderr=log_file)
+                subprocess.call([thirdparty_binary('ivector-extractor-init'),
+                                 '--ivector-dim=' + str(self.ivector_dimension),
+                                 '--use-weights={}'.format(self.use_weights).lower(),
+                                 full_ubm_path,
+                                 init_ie_path],
+                                stderr=log_file)
+            else:
+                subprocess.call([thirdparty_binary('gmm-global-to-fgmm'),
+                                 diag_ubm_path,
+                                 full_ubm_path],
+                                stderr=log_file)
+                subprocess.call([thirdparty_binary('ivector-extractor-init'),
+                                 '--ivector-dim=' + str(self.ivector_dimension),
+                                 '--use-weights=false',
+                                 full_ubm_path,
+                                 init_ie_path],
+                                stderr=log_file)
+
         # Do Gaussian selection and posterior extraction
-        gmm_gselect('final', self, self.corpus.num_jobs)
-        gauss_to_post(self, self.corpus.num_jobs)
+        if self.use_vad:
+            gmm_gselect('final', self, self.corpus.num_jobs, vad=self.use_vad)
+        gauss_to_post(self, self.corpus.num_jobs, vad=self.use_vad)
         parse_logs(log_directory)
         self.logger.info('Initialization complete!')
 
@@ -263,29 +308,77 @@ class IvectorExtractorTrainer(BaseTrainer):
         else:
             iters = range(0, self.num_iterations)
         try:
-            for i in iters:
-                # Accumulate stats and sum
-                acc_ivector_stats(self, self.corpus.num_jobs, i)
+            log_dir = os.path.join(self.train_directory, 'log')
+            if not os.path.exists(os.path.join(self.train_directory, 'final.ie')):
+                for i in iters:
+                    # Accumulate stats and sum
+                    acc_ivector_stats(self, self.corpus.num_jobs, i, vad=self.use_vad)
 
-                # Est extractor
-                log_path = os.path.join(self.train_directory, 'log', 'update.{}.log'.format(i))
-                with open(log_path, 'w') as log_file:
-                    extractor_est_proc = subprocess.Popen([thirdparty_binary('ivector-extractor-est'),
-                                                           '--num-threads={}'.format(self.corpus.num_jobs),
-                                                           '--gaussian-min-count={}'.format(self.gaussian_min_count),
-                                                           os.path.join(self.train_directory, '{}.ie'.format(i)),
-                                                           os.path.join(self.train_directory, 'acc.{}'.format(i)),
-                                                           os.path.join(self.train_directory, '{}.ie'.format(i + 1))],
-                                                          stderr=log_file)
-                    extractor_est_proc.communicate()
+                    # Est extractor
+                    log_path = os.path.join(log_dir, 'update.{}.log'.format(i))
+                    with open(log_path, 'w') as log_file:
+                        extractor_est_proc = subprocess.Popen([thirdparty_binary('ivector-extractor-est'),
+                                                               '--num-threads={}'.format(self.corpus.num_jobs),
+                                                               '--gaussian-min-count={}'.format(self.gaussian_min_count),
+                                                               os.path.join(self.train_directory, '{}.ie'.format(i)),
+                                                               os.path.join(self.train_directory, 'acc.{}'.format(i)),
+                                                               os.path.join(self.train_directory, '{}.ie'.format(i + 1))],
+                                                              stderr=log_file)
+                        extractor_est_proc.communicate()
+
                 # Rename to final
-            shutil.copy(os.path.join(self.train_directory, '{}.ie'.format(self.num_iterations)),
-                        os.path.join(self.train_directory, 'final.ie'))
-            os.makedirs(self.corpus.ivector_directory, exist_ok=True)
-            shutil.copy(os.path.join(self.train_directory, 'final.ie'),
-                        os.path.join(self.corpus.ivector_directory, 'final.ie'))
-            shutil.copy(os.path.join(self.train_directory, 'final.dubm'),
-                        os.path.join(self.corpus.ivector_directory, 'final.dubm'))
+                shutil.copy(os.path.join(self.train_directory, '{}.ie'.format(self.num_iterations)),
+                            os.path.join(self.train_directory, 'final.ie'))
+            # Sliding window CMVN over segments
+            if self.use_vad:
+                self.feature_config.generate_ivector_extract_features(self.corpus, apply_cmn=self.apply_cmn, logger=self.logger)
+            extract_ivectors(self.train_directory, self.corpus.split_directory(), self, self.corpus.num_jobs, vad=self.use_vad)
+            ivector_path = os.path.join(self.train_directory, 'ivectors.scp')
+            plda_path = os.path.join(self.train_directory, 'plda')
+            mean_path = os.path.join(self.train_directory, 'mean.vec')
+            transform_path = os.path.join(self.train_directory, 'trans.mat')
+            if self.use_vad:
+                with open(os.path.join(log_dir, 'mean.log'), 'w', encoding='utf8') as log_file:
+                    mean_proc = subprocess.Popen([thirdparty_binary('ivector-mean'), 'scp:'+ivector_path, mean_path], stderr=log_file)
+                    mean_proc.communicate()
+                    est_proc = subprocess.Popen([thirdparty_binary('est-pca'),
+                                                 '--read-vectors=true', '--normalize-mean=false',
+                                                 '--normalize-variance=true',
+                                                 '--dim={}'.format(self.pca_dimension),
+                                                 'scp:'+ivector_path, transform_path], stderr=log_file)
+                    est_proc.communicate()
+                subsegment_dir = os.path.join(self.corpus.split_directory(), 'subsegments')
+                with open(os.path.join(log_dir, 'plda.log'), 'w', encoding='utf8') as log_file:
+                    compute_plda_proc = subprocess.Popen([thirdparty_binary('ivector-compute-plda'),
+                                                          'ark:' + os.path.join(subsegment_dir, 'spk2utt'),
+                                                          'scp:'+ ivector_path,
+                                                          plda_path],
+                                                         stderr=log_file)
+                    compute_plda_proc.communicate()
+                parse_logs(log_dir)
+            else:
+                x = []
+                y = []
+                speakers = sorted(self.corpus.speak_utt_mapping.keys())
+                for i in range(self.corpus.num_jobs):
+                    ivec = load_scp(os.path.join(self.train_directory, 'ivectors.{}'.format(i)))
+                    for utt, ivector in ivec.items():
+                        ivector = [float(x) for x in ivector]
+                        s = self.corpus.utt_speak_mapping[utt]
+                        s_ind = speakers.index(s)
+                        y.append(s_ind)
+                        x.append(ivector)
+                x = np.array(x)
+                y = np.array(y)
+                clf = GaussianNB()
+                clf.fit(x, y)
+                clf_param_path = os.path.join(self.train_directory, 'speaker_classifier.mdl')
+                dump(clf, clf_param_path)
+                classes_path = os.path.join(self.train_directory, 'speaker_labels.txt')
+                with open(classes_path, 'w', encoding='utf8') as f:
+                    for i, s in enumerate(speakers):
+                        f.write('{} {}\n'.format(s, i))
+
         except Exception as e:
             with open(dirty_path, 'w'):
                 pass
