+++ setup.py
@@ -68,6 +68,7 @@ if __name__ == '__main__':
                     'montreal_forced_aligner.g2p',
                     'montreal_forced_aligner.gui',
                     'montreal_forced_aligner.lm',
+                    'montreal_forced_aligner.multiprocessing',
                     'montreal_forced_aligner.thirdparty',
                     'montreal_forced_aligner.trainers'],
           install_requires=[
@@ -79,6 +80,8 @@ if __name__ == '__main__':
               'pyqt5',
               'pyqtgraph',
               'requests',
+              'sklearn',
+              'joblib'
           ],
           python_requires='>=3.8',
           entry_points={
