--- tests/data/configs/basic_train_ivector.yaml
@@ -1,16 +1,0 @@
-use_mp: false
-
-features:
-  type: "mfcc"
-  use_energy: false
-  frame_shift: 10
-  pitch: false
-
-training:
-  - ivector:
-      num_iterations: 2
-      gaussian_min_count: 2
-      ubm_num_iterations_init: 4
-      silence_weight: 0.0
-      posterior_scale: 0.1
-      max_count: 100
