+++ tests/data/configs/mono_train.yaml
@@ -4,7 +4,7 @@ use_mp: false
 
 features:
   type: "mfcc"
-  use_energy: false
+  use_energy: true
   frame_shift: 10
   pitch: false
